Focus on an item whenever a deep link is clicked or when a page loads
